requirements.txt
langchain==0.1.12
langchain-openai
python-dotenv==1.0.1
python-telegram-bot==21.4
fastapi
databases
sqlalchemy
psycopg2-binary
requests
openai==1.61.1
pydub
uvicorn
asyncpg
websockets
boto3==1.34.47
pandas
matplotlib
seaborn
tqdm
jupyter
ipykernel